app.UseCors(CorsOptions.AllowAll); 加到全域以解決 三種Cors問題如下：
Access to XMLHttpRequest at 'http://localhost:44369/api/Poll/GetVoteResults?Id=1'
from origin 'http://localhost:8080' has been blocked by CORS policy:
The value of the 'Access-Control-Allow-Origin' header in the response must not be the wildcard '*' when the request's credentials mode is 'include'.
The credentials mode of requests initiated by the XMLHttpRequest is controlled by the withCredentials attribute.

Access to XMLHttpRequest at
from origin  has been blocked by CORS policy: The value of the 'Access-Control-Allow-Credentials' header in the response is '' which must be 'true' when the request's credentials mode is 'include'. T
he credentials mode of requests initiated by the XMLHttpRequest is controlled by the withCredentials attribute.

Access to XMLHttpRequest at 'http://localhost:44369/api/Poll/GetVoteResults?Id=1' from origin 'http://localhost:8080' has been blocked by CORS policy:
The value of the 'Access-Control-Allow-Credentials' header in the response is '' which must be 'true' when the request's credentials mode is 'include'.
The credentials mode of requests initiated by the XMLHttpRequest is controlled by the withCredentials attribute.